+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * The MQTT client does not own the miniconf settings struct anymore.
 * The MQTT settings validation handler now receives the already updated new settings
   and the old settings (instead of the not updated current settings and the new settings).
+* The MQTT client has been split into its own `miniconf_mqtt` crate.
 
 ### Added
