+++ miniconf/Cargo.toml
@@ -17,30 +17,8 @@ serde = { version = "1.0.120", default-features = false }
 miniconf_derive = { path = "../miniconf_derive" , version = "0.9" }
 itoa = "1.0.4"
 serde-json-core = { version = "0.5.1" , optional = true }
-log = {version = "0.4", optional = true }
-heapless = { version = "0.8", features = ["serde"], optional=true }
-minimq = { version = "0.8.0", optional = true }
-smlang = { version = "0.6", optional = true }
-embedded-io = { version = "0.6", optional = true }
 
 [features]
-default = ["mqtt-client"]
+default = ["json-core"]
 json-core = ["dep:serde-json-core"]
-mqtt-client = ["json-core", "dep:minimq", "dep:smlang", "dep:log", "dep:heapless", "dep:embedded-io"]
 std = []
-
-[dev-dependencies]
-machine = "0.3"
-env_logger = "0.10"
-std-embedded-nal = "0.2"
-tokio = { version = "1.9", features = ["rt-multi-thread", "time", "macros"] }
-std-embedded-time = "0.1"
-serde = { version = "1.0.120", features = ["derive"], default-features = false }
-
-[[example]]
-name = "mqtt"
-required-features = ["mqtt-client"]
-
-[[example]]
-name = "readback"
-required-features = ["json-core"]
