+++ miniconf_mqtt/Cargo.toml
@@ -13,3 +13,19 @@ categories = ["no-std", "config", "rust-patterns", "parsing"]
 
 [dependencies]
 miniconf = { version = "0.9", path = "../miniconf" }
+minimq = { version = "0.8.0"}
+smlang = { version = "0.6"}
+embedded-io = { version = "0.6"}
+log = { version = "0.4" }
+heapless = { version = "0.8", features = ["serde"] }
+
+[[example]]
+name = "mqtt"
+
+[dev-dependencies]
+machine = "0.3"
+env_logger = "0.10"
+std-embedded-nal = "0.2"
+tokio = { version = "1.9", features = ["rt-multi-thread", "time", "macros"] }
+std-embedded-time = "0.1"
+serde = { version = "1.0.120", features = ["derive"], default-features = false }
