+++ miniconf_mqtt/tests/integration_test.rs
@@ -101,8 +101,8 @@ fn main() -> std::io::Result<()> {
     let localhost: minimq::embedded_nal::IpAddr = "127.0.0.1".parse().unwrap();
 
     // Construct a settings configuration interface.
-    let mut interface: miniconf::MqttClient<'_, _, _, _, minimq::broker::IpBroker, 2> =
-        miniconf::MqttClient::new(
+    let mut interface: miniconf_mqtt::MqttClient<'_, _, _, _, minimq::broker::IpBroker, 2> =
+        miniconf_mqtt::MqttClient::new(
             Stack,
             "device",
             StandardClock::default(),
